requirements.txt
discord.py==2.3.2
pymongo==4.6.1
python-dotenv==1.0.1
motor==3.5.1
dnspython==2.6.1
apscheduler==3.10.4